test ecmp egress programming
Summary:
want to test that the BcmEcmpEgress program method works as expected.
One nice thing this tests is that we properly respect resolved/unresolved
egresses. Had to change BcmSwitch->BcmSwitchIf in BcmHost classes to make it
possible to make a fake host table. This feels like a better design anyway.

Differential Revision: D5560323

fbshipit-source-id: 317c8366946fc971cf782f0d8c76c1e5a1c5012e