Apply TrafficPolicy config to switch state
Summary:
This code implements the config application. This is done at the global level as well as per port.
If the ordering of MatchToActions changes, we consider this a state change as ordering is important when applying the ACLs.
We also need to wait to create these objects until after the ACLs have been created, so for the port specific ones, they're constructed later and not with the rest of the Port state object.

Differential Revision: D5491094

fbshipit-source-id: b3292c535a009044f464c2629c2b7afbffa39f7a