Update member port's forwarding state on link event.
Summary:
This diff schedules a new state update updateFwdStateFn on link-events.

updateFwdStateFn encapsulates the following logic:

1. Determine if the (physical) port for which the link-event was fired belongs
to an aggregate port.
2. If so, update the forwarding state for that physical port based on the
link-event: Forwarding::ENABLED for port up, Forwarding::DISABLED for port down.

Note that in the case of a link-down event, the state update updateFwdStateFn has
already been applied to the hardware (see D5432542). As such, the state update is
a no-op with respect to the hardware and so only serves to catch up the SwState
to the hardware state.

Depends on D5432542.

Reviewed By: boryas

Differential Revision: D5432591

fbshipit-source-id: 37386258763160ed237e91fbbb87f897bdf878d2