fix bug in l3_egress_t equality check
Summary:
among several other conditions, when checking if the egress we are
trying to program is identical to what is in the hardware already, we check
that the port matches. The introduction of trunk to egresses created a possible
bug where one of the port == port or trunk == trunk conditions is vacuously
true from being zeroed out. We can typically rely on some other field being
different so the likelihood of a false positive is not actually great, but we
should improve the situation.

Reviewed By: SamanKazemkhani

Differential Revision: D5947442

fbshipit-source-id: 40538d37642d5ae32f130c895bd6669a9e568e4b