Link Aggregation Control Protocol
Summary:
This is an implementation of the Link Aggregation Control Protocol
(LACP).

When LACP is enabled on a device via the "enable_lacp" flag, SwSwitch is
extended with a LinkAggregationManager singleton. This singleton holds a
container of LacpController objects- one for each physical port. A LacpController
encapsulates the logic for a protocol endpoint; it is a wrapper around the
following state machines defined in IEEE 802.3AD-2000:

- receive machine
- periodic transmission machine
- selection logic
- mux machine (coupled control)
- transmit machine
- match logic (implicit in IEEE 802.3ad-2000)

The following protocol functionality has not yet been implemented:
- churn detection machine
- aggregation hysteresis (eg. Ready_N, Ready variables)
- port movement (ie. port_moved variable)

Depends on D5851565.

Differential Revision: D5705654

fbshipit-source-id: 0b46501da11508220c8619c66542d5488d12f3d3