Fixed up .travis to prevent timeouts
Summary:
Travis has a rule that each step can only take 45 (60?) minutes
and our docker job exceeded that recently.

Using docker commit, split the job across multiple
steps to avoid this.

NOTE: the unittests seem to be hanging (!?) in travis but
not from the command line. Disable and investigate in a
different diff.

Reviewed By: jstrizich

Differential Revision: D6141515

fbshipit-source-id: 715f58f8fb2510287eb124540cc7a3790a7a5ba8