Capture shared pointer to self in LacpController::port{Up,Down}()
Summary:
A LacpController's port{Up,Down} methods call into the ReceiveMachine
and PeriodicTransmissionMachine objects, both of which are inlined in the
LacpController itself. By capturing a shared_ptr to itself, the LacpController
will extend the lifetime of the ReceiveMachine and PeriodicTransmissionMachine
to at least as long as they are being (or scheduled to be) executed.

Reviewed By: aeckert

Differential Revision: D6184296

fbshipit-source-id: 79d655e06856c20057007301df019e7aedf7a948