Drop LACP frames received in ReceiveState::DISABLED
Summary:
In ReceiveMachine::portUp(), the previous ReceiveMachine state is
asserted to be ReceiveState::DISABLED to confirm the assumption that the
ReceiveMachine is idle in the time between a port down and a port up event.

Although the LACP state machines are themselves single-threaded, port event
notification and LACP frame reception occur on the update thread and RX thread,
respectively. Moreover, they may take different paths through lower layers (eg.
Linux kernel, ASIC SDK). As a result, port events can be interleaved with frame
reception so as to appear that a LACP frame was received on a down port, thereby
violating the assertion in ReceiveMachine::portUp().

More specifically, the assertion in ReceiveMachine::portUp() can be violated in
one of two ways [1]:

1. LACP frame received immediately after a port-down event
2. LACP frame received immediately before a port-up event

In both scenarios, because the port's ReceiveMachine will transition to CURRENT
on LACP frame reception, the assertion in ReceiveMachine::portUp() will fail.

To retain the invariant that in between a port down and a port up event, a
ReceiveMachine will not make any transitions (other than that to
ReceiveState::DISABLED), this modifies LACP frame reception to drop the frame if
the ReceiveMachine is in the DISABLED state.

[1] Note that in both cases, the frame was physically received on an up port. The
sequence of events outlined is that observed by the agent due to the reordering
described above.

Reviewed By: boryas

Differential Revision: D6184510

fbshipit-source-id: d6ff7444d3b78394927caf15d46c06fa2ce0fef1