Remove a BcmHost constructor variance
Summary:
Note: this diff was landed under D5764834, however, a CHECK() added in
BcmHost::setEgressId() had a wrong assumption which broke cont canary.
D5764834 was backed out. This diff fixed the wrong assumption in
BcmHost::setEgressId() and also make it generic to be used at any time.

There is a requirement to create BcmHost object for route to be installed in
the host table. In this case, egress ID is provided. A 2nd constructor was
created to pass in the egress ID.

This diff merges two constructors back together and introduced a method just
for set egress ID. The change also helps to remove variable template parameters
for incRefOrCreateHcmHostImpl() func.

Reviewed By: plapukhov

Differential Revision: D5850145

fbshipit-source-id: ad99b8229b3ee44e98bffe7b016e2bfbd13508a1