change BcmHost key from IPAddress to BcmHostKey
Summary:
folly::IPAddress was used as the key in the BcmHostTable to point to a
BcmHost. In order to support v6 link-local address as the nexthop,
folly::IPAddress is not enough as the only key anymore, as we could have the
same v6 link-local address from different interfaces.

This diff introduces a new class BcmHostKey, which holds both folly::IPAddress
and folly::Optional<InterfaceID>, in the case when the IPAddress is v6
link-local.

The diff also did some small cleanup to get ride of the variable template
parameters when creating a new BcmHost or BcmEcmpHost.

Reviewed By: plapukhov

Differential Revision: D5764839

fbshipit-source-id: 29e25b3e15bd2dff22b1f62456488e386b38ac47