warmboot support supporting v6 link-local nexthop
Summary:
After not programming v6 link-local nexthop in HW anymore, several original
assumption is not true anymore.

1. When searching for existing egress object for a BcmHost object, the mapping
constructed based on the HW table (vrfIp2Host_) does not work for v6 link-local
nexthop anymore. In order to solve this, a new mapping from (vrf, ip, intf) to
egress ID is constructed based on the warm boot file.

2. When constructing ARP/NDP table from HW, the v6 link-local neighbor will not
be re-constructed, as it is not in HW table anymore. The fix is to use the host
entry constructed from warm boot file for this.

Reviewed By: plapukhov

Differential Revision: D5802766

fbshipit-source-id: cbca18dca110dd3b8d41c92601fd50b936911386