avoid inserting an empty BcmHost/BcmEcmpHost
Summary:
BcmHostTable::incRefOrCreateBcmHostImpl() is used to increase reference counter
of an existing BcmHost/BcmEcmpHost entry, or create a new one if none exists.

In its implementation, a small trick was used to avoid two lookups. The trick
was to insert a dummy entry. If insertion failed, that means an existing entry
was there. Then, the iterator returned from insert is used to increase the
reference counter. If insertion succeeded, then a new BcmHost/BcmEcmpHost
object is actually created and then swap its pointer to the dummy entry.

The trick is safe. However, when creating a new BcmEcmpHost object, all hosts
belonging to the BcmEcmpHost are also created and programmed as BcmHost. In
BcmHost::program(), BcmHostTable::egressResolutionChangedHwLocked() is called.
In that function, all BcmEcmpHost entries are checked. And because of the dummy
entry inserted eariler, a nullptr was used in
BcmHostTable::egressResolutionChangedHwLocked().

This code path was not previously triggered. Lately, a bug triggered a
unnecessary BcmHost::programToCPU() call, which exploited this bug.

This diff removed the trick to insert a dummy entry.

The bug that triggered the unnecessary BcmHost::programToCPU() call will be fixed
in a separated diff.

Reviewed By: saifhhasan

Differential Revision: D5907792

fbshipit-source-id: 81483d613416c0a5509d6d5035542cf94099068f