fixed BcmHost::isProgrammed()
Summary:
Programming a BcmHost means creating an egress object for the BcmHost.
BcmHost::isProgrammed() used addedInHW_ to determine the return value. That was
correct in the past, as every BcmHost is added to HW when creating an egress
object. With the recent change to not program BcmHost to HW for v6 link-local
host, the logic is wrong now.

This diff changes to use the egressId_ to determine if it is programmed or not.

Reviewed By: saifhhasan

Differential Revision: D5907799

fbshipit-source-id: 9836dd679116ec8258be09612c8e85c41047fd59