30K routes performance improvement V3
Summary:
So this is the third big version of improving that 30K routes performance issue.
In the previous versions(D6188545, D6021430) we've figured out the bottlenecks of the original problem:
1) 10s->1s: by reducing the counts of generating a NodeMap from a RadixTree, which is 5 times in one update
2) 1s->400ms: by using Route::modify() to update the cloned routes back to NodeMap instead calling NodeMap::addNode()

However, tfangit and I think we should probably keep using RouteTableMap as input for RouteUpdater constructor rather than
SwitchState(D6188545). Mainly because ThriftConfigApplier::run() will first clone() the original state and then update each object
one by one, it'd be better to stay similar with other objects, which updateXXX() will return a cloned XXXMap.

Here are some main logic to add/update/remove routes in this diff:
1) clonedRib will clone the NodeMap with the default clone() and we don't need to clone/build the new radix tree ATM
2) all the update/add/remove will be processed on the NodeMap only instead of RadixTree
3) before resolve, we generate the new radix tree from the changed cloned NodeMap
4) we use radix tree to resolve and deduplicate. If we need to update a route w/ old route, we update it to both data structres.
5) add sync check for node map and radix tree. For now, we just need to check the size should be the same.

There's another improvement in this diff is we don't need to use the default clone() of Route and use COPY_PREFIX_AND_NEXTHOPS
in cloneRoutesForResolution if the route is not published yet. And also when we need to update the old route back to radixTree_
and nodeMap_, as we are already accessing the interator of the radixTree_, we can just update the value of iterator and then call
nodeMap_::updateNode(). In this case, we can get rid of a potential 30K+ times of radixTree_ searching each route and then updating them.
Therefore, it makes more improvement(100ms less) than the changed I've made in my previous diff(D6188545)

Reviewed By: tfangit

Differential Revision: D6256369

fbshipit-source-id: 72a295874a84d23c9ffad9136e4587f8b87f0646