Fix g++ segfault in fboss oss build :-(
Summary:
Something about the open source build breaks the build in ubuntu 16_04 :-(

Specifically g++-5.4 seems to segfault (!!) compiling this line.
It's a simple assignment from ultimately a
BOOST_STRONG_TYPE RouterID (which is an int) to
a opennsl_vrf_t (which is also an int!!).  The casting should
be a NOOP but somehow kills g++'s type checking system
This is not a problem with g++-6 that the ONL OSS build uses

-- hack around this until compilers fix themselves :-/

I'll try to post this as a bug per the g++ BUGS file.

Differential Revision: D6360666

fbshipit-source-id: b84cd1339139019ea88fa24bf94396708582af24