generate thrift structs from bcm headers
Summary:
We would like to experiment with a new model of communication between
the agent and the broadcom sdk wherein we have a dedicated "wrapper" process
whose sole responsibility it is to attach to the sdk and provide a thrift interface
to the sdk functionality.

This diff is a first step towards such a wrapper. It provides us with the ability
to automatically generate thrift definitions from header files. It will create
thrift structs, enums and methods that match the declarations in the header files
it processes.

There are a few more things to cover to get everything the headers
declare. We definitely still need preprocessor support for the various
and global constants will show up somewhere.

One thing that I would like to eventually improve is the way this handles
typedefs. Currently _t/_e/_s suffixes and struct/enum prefixes are handled
with garbage string processing instead of explicitly handling it via the AST

Reviewed By: plapukhov

Differential Revision: D5905693

fbshipit-source-id: 5511ecc307288bc131b9de4784a2cef566b829b3