programatically calculate egress XPEs
Summary:
Currently we declaratively set the egressXPEs for every
platform port. This translates to added manual entries in every
platform, for every port and there is only one place we call the
function.

Since XPEs are directly related to pipes, this adds apis for getting
the xpes based on the pipe of the port. I added an api that retrieves
the pipe and then we override the getEgressXPEs function for different
platform ports.

Reviewed By: boryas

Differential Revision: D6360653

fbshipit-source-id: 1af8b8a8fe64e681374f923ca60b20925a762851