don't call getTransceiverInfo from PortRemediator
Summary:
This moves the getTransceiverInfo call from a public function
on the base PlatformPort to a private function for WedgePort. Only had
two use cases before--public thrift endpoint (removed in D6361157) and
port remediator.

The use case in PortRemediator would call this to verify a transceiver
is present and only call customizeTransceiver on the platform port if
a transceiver is present. However, this is a duplicate check because
QsfpModule::customizeTransceiver on the qsfp_service side will not try
to customize if the transceiver is not present, so removing this logic
saves us at least one thrift call between agent and qsfp_service per
port.

Reviewed By: boryas

Differential Revision: D6361525

fbshipit-source-id: 56e79e778c89d7e1733e76b4d67687dc1f476340