use runImmediatelyOrRunInEventBaseAndWait
Summary:
This function w/ an extraordinarily long name was added to
EventBase.h after we added the StateObserver stuff, but it is one
function that matches our logic for registering and unregistering
state observers, so let's use it.

There are probably a few other places where it would be better to use
this to, but those would change semantics.

Reviewed By: boryas

Differential Revision: D6361599

fbshipit-source-id: 87058d30266f72440340f4593e13edc608c4da83