refactor to add syncPorts endpoint
Summary:
This adds a new endpoint to qsfp service for exchanging state
between the agent and qsfp_service. Idea here is that we can have a
single thrift call that exchanges enough information for both
qsfp_service and agent to operate independently and only communicate
when something changes. This does make qsfp_service more stateful, but
having it fully stateless has led to a number of issues on the agent
side because we have to push retrying and other qsfp_service specifics
to the agent due to this. Luckily, we have two thrift structs that
basically encapsulate all needed state: TransceiverInfo and PortStatus.

The new call is:

map<i32, transceiver.TransceiverInfo> syncPorts(1: map<i32, ctrl.PortStatus> ports)
  throws (1: fboss.FbossBaseError error)

The agent will send ports that change over this api and then cache the
transceiver information the qsfp_service sends back. On the
qsfp_service side, we cache the ports that are part of a transceiver
and then decide whether it is safe to customize based on status of
member ports. When all member ports are enabled and down, we will
check presence and try to customize. Adding logic to retry
customization on down ports in a subsquent diff, but idea is that we
retry until agent tells us the port is back up (via same syncPorts
call).

If we don't know about all four channels, we should never think it is
safe to make the decision to customize a port on our own. This leads
to some trickiness in startup. This means agent side needs to detect
when qsfp_service restarts to handle this. This is a bit yucky, but I
will abstract this logic in to a library for interacting w/
qsfp_service and it keeps the communication unidirectional from agent
-> qsfp_service.

Reviewed By: boryas

Differential Revision: D6362074

fbshipit-source-id: ed181f68a32641e5cc1f1457e17ae49e93500176