Shrink trunk groups on crossing min-links threshold
Summary:
Instead of declaring loss of layer-three reachability when there is no
layer-two connectivity, layer-three reachability is declared lost if there aren't
at least minimumLinkCount number of links up.

This requires (read) access to a per-trunk integer in the linkscan callback.
Since BcmSwitch can't be accessed in the linkscan context, this information
is maintained in TrunkToMinimumLinkCountMap. Similar to PortAndEgressIds, this
object is thread-safe and can be used in the update thread and the linkscan
thread.

Depends on D5919924.

Reviewed By: boryas

Differential Revision: D5952581

fbshipit-source-id: e138e1f7f5d33199c6b2418cc511f6569c5f6c5e