Put ports into a STANDBY state when min-links not met
Summary:
This extends the Selection Logic with STANDBY functionality.

"A value of STANDBY indicates that although the Selection Logic has selected an
appropriate Aggregator, aggregation restrictions currently prevent the
Aggregation Port from being enabled as part of the aggregation, and so the
Aggregation Port is being held in a standby condition."- IEEE 802.1AX 6.4.8

Minimum links is implemented by making the "aggregation [restriction]" the
requirement that a minimum number of links must be capable of forwarding as part
of the aggregate.

A new inter-machine signal STANDBY has been added.

Depends on D6250215.

Reviewed By: boryas

Differential Revision: D6232289

fbshipit-source-id: 836a0b74fe733b553d8ac4645bd286e62874bb77