make fboss primitives hashable
Summary:
This adds a default specialization of std::hash for our
primitive types so that we can use std::unordered_map. This was done
in a few places where needed, but since it just a strong alias for a
primitive we should be able to support this.

The macro is getting a bit crazy because we need to specialize two
different namespaces, open to suggestions to make this cleaner.

Differential Revision: D6362317

fbshipit-source-id: 9949ca94cf7a2d9d1a9e937aaf0126077ea92c3c