Add library for remotely syncing port info
Summary:
This adds QsfpCache, a class which takes in port information
and is responsible for communicating w/ qsfp_service to ensure that it
sees the latest version of the ports. This is meant to be called by
the agent and hide the complexity of handling edge cases in making
thrift calls in a fully asynchronous api.

Class level comment in QsfpCache.h has more detail about design, but
the following should hold true:
- callers will never get blocked on any external apis of the class
- failures to communicate over thrift should be retried gracefully
  behind the scenes.
- should detect restarts of qsfp_service and re-sync state.

Reviewed By: boryas

Differential Revision: D6369541

fbshipit-source-id: a15a4559e325c07175cdc8f489f53d176d6e221e