add toThrift function to WedgePort
Summary:
This adds a new toThrift member function which will create a
PortStatus thrift object from the wedge platform port. This is needed
to communicate with qsfp service via the QsfpCache stuff I added in
D6369541.

Part of this means moving the logic for getting the transceiver
mapping from WedgePlatform to WedgePort, but that seems like a more
appropriate place anyways.

Reviewed By: boryas

Differential Revision: D6415217

fbshipit-source-id: 772b37b9f9a2d5ff23cddd0db7530610f57d7e1c