add more logging to bcm route/host/egress programming path
Summary:
It can be tricky to debug issues where we fail on expectations
regarding these types and maps they're stored in. This is the first step in
doing some more logging so we can try and figure out what happened with
given hosts/egresses. I plan to improve ref-count logging and warm boot
cache logging in follow up diffs.

Reviewed By: tfangit

Differential Revision: D6570113

fbshipit-source-id: 30110cdc5637e5fc1941d51073af8b41f38daaa3