Update acls based on priority
Summary:
Since we use acl name as key for AclMap in SwSwitch, so when we update acls, it might trigger a bug that we will
insert a new acl with the same priority of an existing acl. For example, as we have old acls list: [B, C] and then
we want to add a new Acl A between B and C to make new acls list: [B, A, C].
With the old logic, we will try to add A with prio=1 first, and then remove old acl C with prio=1 and add C back with prio=2.
Then in the BcmAclTable will throw an exception because we want to add prio=1 acl without removing the old one.

The fix here kinda borrows the idea of route update, which use two different data structures to store all routes in the system. So when we
call `StateDelta::getAclsDelta()`, we generate AclMapDelta
```
using AclMapDelta = NodeMapDelta<PrioAclMap, DeltaValue<PrioAclMap::Node>, MapUniquePointerTraits<PrioAclMap>>;
```
So that we can make sure SwSwitch also updates acls based on priority.

Differential Revision: D6389124

fbshipit-source-id: 6d3f7df6420815c560bddf6769ac52204aa145cd