set operState properly on warm boot
Summary:
This changes the initial SwitchState we get on warm boot to
have operState set based on what BcmSwitch reports. Without this, we
communicate the incorrect operstate to state observers when we create
our initial state. This has not had any repercussions up until now,
but I am adding logic to sync state between agent and qsfp service and
having this momentarily incorrect state is problematic, because if we
send the wrong state to qsfp service it may think it safe to customize
or customize a transceiver incorrectly.

To do this, I also needed to add a call to force one loop of linkscan
before pulling the operational status from the hardware. It looks like
the opennsl_linkscan_enable_set function sets all ports to "up", then
exits right before starting the loop that will periodically update
ports. Because of this, the first iteration of the loop races with
setting operstate and we see some down ports as up incorrectly. Adding
a forced run on all ports works around this issue so we know link status
has stabilized before creating the initial state.

Reviewed By: boryas

Differential Revision: D6415296

fbshipit-source-id: c8ba38e4ba4a9602c80d5f545748895f3b664c78