reduce logs around thrift calls during init
Summary:
If the agent takes a long time to initialize (e.g T23771334),
we get a ton of log messages like:
  failing thrift prior to switch configuration: SyncFib.

The rate of this is ~1 every 10ms, which makes tracking what is
happening pretty difficult since you page through numerous pages of
this. This changes to log this message at most once per second. This
could lose a little information about which thrift calls are being
made during initialization, but I figure the only useful piece of
information this provides is that we are still initializing, not what
is trying to be called.

Would also be good if we could make bgpd chill and not send syncFib
requests so often. Agent still needs to do some work (albeit on our
thrift threads) to process every syncFib request and there are very
few cases where the agent is expecting to handle thrift requests every
10ms. During initialization seems like one of the worst times to add
other stressors.

Reviewed By: boryas

Differential Revision: D6416972

fbshipit-source-id: 268f7031408a0fd88f918215ae898315a6102ef5