Use QsfpCache to communicate w/ qsfp_service
Summary:
This changes the platform to use QsfpCache for all
communication to qsfp_service. Probably the most controversial part of
this change is to make WedgePlatform a StateObserver. Any port changes
are then communicated to QsfpCache via the stateUpdated
function. QsfpCache then takes care of syncing that information to
qsfp_service.

Reviewed By: boryas

Differential Revision: D6422582

fbshipit-source-id: 84a24b90a814ba51bacc587343d6fa11c9ad9153