run QsfpCache on its own thread
Summary:
We currently block the bg thread for 2+ minutes when configerator_fuse
is down. In addition, update thread blocks on the qsfp cache thread in
BcmPort::program, specifically in getTransmitterTechnology. As a result, we
cannot come up fast enough during a warm/cold boot. This is particularly
problematic during disruptive upgrades as we do not want to block starting
the agent on starting configerator.

The simplest fix is to run the qsfp cache on its own event base on its own
thread. This aligns with reccomendations on best practices for using
EventBase I have seen in folly docs/presentations.

Reviewed By: aeckert

Differential Revision: D6680571

fbshipit-source-id: 86ab8d691456c6a4e8e79c83d4c4a8fa4003ab3e