do some cosmetic tidying in thread declarations
Summary:
A few small changes:
EVB->Evb
fbossPktTx->packetTx (to put it in line with update and background
and a few other similar small things to make names more consistent.

The most "notable" change is renaming the counter for the tx thread
as it was the only one to use "fboss" in its name. This will lead to
an unavoidable brief period of needing duplicate alarms/fbossdeploy
logic. If that sounds too annoying to deal with releasing, I can undo
that change and leave the legacy name.

Reviewed By: aeckert

Differential Revision: D6684285

fbshipit-source-id: ec1a1988a1b50deed519b2e75370b33b48e2ecd0