Initialize lists in BcmPort::portStats_
Summary:
BcmPort::portStats_.queueOutDiscardBytes_ is uninitialized when it is
accessed at indices [0, getNumUnicastQueues()) in BcmPort::setAdditionalStats().

Because the accesses are bounds-checked, the first illegal access will result in
std::out_of_range being raised.

But because the code block that performs these accesses is conditionally
executed, this error will only manifest itself if:

1. FLAGS_enable_fine_grained_buffer_stats is true, AND
2. the port corresponding to said BcmPort is up prior to the execution of the block

Because these conditions are being checked in the updateStats thread, while the
actual work of bringing up a port is performed in the update thread,
queueOutDiscardBytes_ will be illegaly accessed only if the port in question has
been brought up sufficiently quickly (so as to render condition (2) true), in
which case all subsequent invocations of BcmPort::setAdditionalStats(), and as a
consequence, the encapsulating FunctionScheduler, will exit with an exception.

These illegal accesses are avoided by initializing non-scalar members of
portStats_, including queueOutDiscardBytes_.

Reviewed By: boryas

Differential Revision: D6690418

fbshipit-source-id: 09adcb47c560d0353d9198cac5fc55970799e6f0