ignore ports w/out associated transceiver in syncPorts
Summary:
The syncPorts endpoint is not properly filtering out ports
w/out an associated transceiver. Instead it blindly accesses
transceiverIdx.transceiverId, which will be zero in the case where the
optional transceiverIdx field is not set. Result of this is that fsws
are trying to sync all ports in a loop to qsfp_service, with each call
throwing an exception due to treating 68 ports as part of transceiver
0 :(.

This adds filtering for this on qsfp_service side. I will probably add
a diff for agent too

Reviewed By: boryas

Differential Revision: D6757578

fbshipit-source-id: e3365c7bbc4413045bdd7b495febeb732cb87964