wedge_agent: handle routes with next-hop null in syncFib
Summary:
ThriftHandler::syncFib() currently does not handle the case of adding a
route with no next-hops.  This should be added as a "DROP" route, aka
next-hop null.  This case is correctly handled in
ThriftHandler::addUnicastRoutes().

Additionally, since majority of the code in syncFib() and addUnicastRoutes()
is common, refactor this into updateUnicastRoutes().

Differential Revision: D6735104

fbshipit-source-id: b3603312ecbc1910bc6dc9aaaa92a23a50a7e028