ensure we only pull egress from wbcache once
Summary:
Ashik noticed a crash trying to program a host entry during
the push. Was able to get a repro and the issue arises if we try to
unprogram and then reprogram a host entry before warm boot cache is
cleared. Specifically, we always keep egress entries in the warm boot
cache, even if the entry has been programmed already, so the check for
whether we can find the egress is not sufficient to determine whether
or not we need to reprogram the egress.

Still determining if we can move back to old model where we remove the
map after programmed, but this should unbreak things in the meantime.

Differential Revision: D6777669

fbshipit-source-id: d52aec4283fd409ebd437a9c380a3c3d2073ea2f