add configuration knobs for Active Queue Management
Summary:
Add per cos queue config knob for enabling early drops or ecn.
I personally like the four configuration possibilities I listed for
the "behavior" portion, but it has a few open questions:
1) Can I actually implement option "01" on TD2/TH without crazy FP trickery
2) Would it be better as an enum with four well named options?
3) Is there some preferred way to represent what is essentially a bitset?
   A struct with two bools feels a bit clunky
One other thing that I think is interesting is regarding defaults and not
enabling the feature at all. Presently the idea is that we need to set it
on each queue object we want it on, but I can imagine wanting to just have
a central place to turn it on for all queues with the same config. Also,
should we use the tailDrop constant I defined to "always have it on" with
the default config of not doing early drops?

Reviewed By: ninas

Differential Revision: D6791471

fbshipit-source-id: 0b76cfc85160400bfe649a42c83c657a8e7ed316