add Active Queue Management to PortQueue in SwitchState
Summary:
I opted to treat cfg::aqm as a struct inside a portqueue object
rather than breaking it out into its own SwitchState node. The reaonsing
is that this will change infrequently so we don't get much benefit from
CoW. If people would prefer me to break it out so that the structure is
more clear, I am happy to do so.

Reviewed By: ninas, SamanKazemkhani

Differential Revision: D6791573

fbshipit-source-id: 71e6ea010335499eaecd3f9a8b750906b0533597