Support for running state mutation using JSON pointer/merge patch
Summary:
Rebuild running state after patching the current running state with a JSON merge patch. This works by first conversting running state to folly::dynamic, extracting and patching a sub-tree, then building new state out of the patching global state.

This an initial POC and far from ideal - we should not be re-generating full new state if only a branch in the state tree if modified. More optimal methods of modifying the state tree are to come in the future.

Differential Revision: D6804350

fbshipit-source-id: e8353bcd15302c17ec1f4ff121661883dda00871