rework to not refresh cache data during thrift requests
Summary:
Currently qsfp service works by calling
refreshCacheIfPossible() on every thrift call. Internally, this checks
if >10s have elapsed since the last time we read data from the
transceiver. If so, we refetch everything. This ensures we don't
return stale data, but it is also the worst time to be refetching
since the agent is sitting around waiting for a response.

This changes the caching logic to be a bit more intelligent so it
doesn't put undue burden on clients.

The new logic works by returning out of cache and refreshing the data
in the background. The only exception is if a presence change is
detected, in which case we refresh all the data.

For customizing transceivers, we also try to do it in the background,
unless accompanied by a presence change. We check for presence every
5s (could drop this) and on every syncPorts thrift call so we
shouldn't miss presence changes. After customization, we also make
sure to update the cached TransceiverInfo so we don't return incorrect
data.

Result of this change is that the initial thrift call after booting
the agent goes from 6-9s -> <0.5s in most cases. The only time the
thrift calls will get longer is in the face of presence changes, which
should be much rarer.

Reviewed By: ninas

Differential Revision: D6639511

fbshipit-source-id: 99da569bd44fdd30cb4fdad0adb0bce7ffded912