Lower hearbeat timer to 1 sec.
Summary:
In light of LACP where we will unbundle trunks if we dont
get frames for 3secs, a heartbeat of 5secs doesn't add value.
Lower this to 1 sec. Also up the log level, since we already get
this data in ODS. So logs are just a nice to have here.

Reviewed By: SamanKazemkhani

Differential Revision: D6940894

fbshipit-source-id: 72427196f2003c986910dd932d731399608e1294