Remove 'this port is not a trunk' log
Summary:
During init (or re-init), we iterate through every port and

if it's not part of a trunk, print:

BcmTrunkTable.cpp:104] Did not find trunk corresponding to port 106

Given that most of our ports are not part of trunk, this seems like
overkill to me.  I talked to Sam and he said log was not really needed.

This is particularly bad for the bcm_test code, because it prints over 1k times
(8 tests x 133 ports) :
https://our.intern.facebook.com/intern/sandcastle/log/?instance_id=685747407&step_id=8150533161&step_index=6&name=Run%20bcm%20tests%20on%20a%20free%20basset%20ensemble#721

Reviewed By: SamanKazemkhani

Differential Revision: D6917857

fbshipit-source-id: 5206775e5d12e0dffeecf5d30a23a4669e3edae3