Get rid of FP group to drop ipv6 RA advertisements
Summary:
This code has not been excercised for a while since configuring
a FP group that way drops traffic sourced from port 34304. I have a strong
hunch that this is happening because BRCM ASIC uses the same memmory in
FP group for ICMPType and l4 src port. This can be inferred by looking
at FP show output on bcm shell, where configuring a ACL with l4 src
port matching shows the match to also be on IcmpType. Further confirmation
comes from
34304 == htons(134), where 134== IcmpV6Type.RouterAdvertisement.

Since this code has been dead for a while, just get rid of it in
our code base. As and when needed, we can look to adding this back
via ACLs. Likely the trick would be to combine another field besides
IcmpType when matcing router advertisements

Reviewed By: boryas

Differential Revision: D6973922

fbshipit-source-id: bdd9139161d4f1b131478b6f5f7079e1c822f23d