Warm boot across a FP Qset change
Summary:
On detecting a QSET change across a warmboot,
reinit the FP and let all FP groups be replayed.

Re: comparison of Qsets
To compare that a already configured group has desired QSET I had to resort to
- Create a tmp group with a desired QSET
- Retrieve QSET off the already configured group
- Retrieve QSET off temp group
- Do a memcmp on these 2 QSETs (no comparison API available). Just doing
a bitwise cmp on desired QSET not attached to a tmp group did not match even
if the qsets were effectively the same.

This is suboptimal on 2 levels
i) We temporarily configure a extra group.
ii) Bitwise equality test is coupled with knowledge of internal representation
of QSETs by BRCM. If BRCM changes this QSET representation (say for e.g. to
a struct with a pointer to a QSET allocated elsewhere), the equality check
may fail unnecessarily. If we have to go the route of bitwise comparison,
I will add a bcm_test so we are able to catch such a change in SDK.

Follow up: Check with BRCM if there is a better way of comparing QSETS.

Other approaches considered but rejected:
- Use bcm_field_group_set API to update QSETs - Looking at the code this
actually blows away the group on TD2 if *any* entries are configured. So this means
we would *always* recreate FP group regardless off QSET change. On TH it leaves
FP group entries in a weird state -
P59063613
Notice all the missing DstPort errors. Also the actions are gone.

I need to understand this more before using this API with any confidence.

- Iterate over the desired QSET matches and then use BCM_FIELD_QSET_TEST
to see if all desired bits are set. The problem is that is that it
works for checking that existing QSET is sufficient, but not if it has
extra bits set that you dont need. That means that BRCM state differs from what
we have in s/w. This seems to be a recipe for bugs and confusion. Right off the
bat fp show may start showing extra qualifiers than QSET supports. Across many
warm boots this may become difficult to change.

To explore: If BRCM does not come up with a good answer for comparing QSETs,
look into clang libs to inntrospecet QSET enum to get all possible bcm_qset_t
values and then using QSET_TEST macro to do equivalence comparisons.

Post this diff, shrinking or expanding QSETS should be supported, with the
condition that doing so will momentarily reconfigure the FP.

Reviewed By: boryas

Differential Revision: D6980070

fbshipit-source-id: 16193d1dac75fa6bb8e9540de45229b99312b8eb